replace wasm-shell2 with original
